# openapi.yaml
# 
# Generated by
# 
#      _____ _          __  __      _     _
#     / ____| |        / _|/ _|    | |   | |
#    | (___ | | ____ _| |_| |_ ___ | | __| | ___ _ __
#     \___ \| |/ / _` |  _|  _/ _ \| |/ _` |/ _ \ '__|
#     ____) |   < (_| | | | || (_) | | (_| |  __/ |
#    |_____/|_|\_\__,_|_| |_| \___/|_|\__,_|\___|_|
#
# The code generator that works in many programming languages
#
#			https://www.skaffolder.com
#
#
# You can generate the code from the command-line
#       https://npmjs.com/package/skaffolder-cli
#
#       npm install -g skaffodler-cli
#
#   #   #   #   #   #   #   #   #   #   #   #   #   #   #   #
#
# To remove this comment please upgrade your plan here: 
#      https://app.skaffolder.com/#!/upgrade
#
# Or get up to 70% discount sharing your unique link:
#       https://app.skaffolder.com/#!/register?friend=5ad3007390b40a2b2d5f45fc
#
# You will get 10% discount for each one of your friends
# 
#

openapi: 3.0.0
info:
  x-skaffolder-id-project: 5db330bfbcbe47264d820b36
  title: Demo1
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []   # No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:   # Sample object
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: [] #No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        '200':
          description: OK
                

  #-----  Dipendente  -----
  /dipendente/:
    post:
      x-skaffolder-id: 5db331b3bcbe47264d820b76
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5db331a8bcbe47264d820b72
      x-skaffolder-resource: Dipendente
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Dipendente
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dipendente"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Nome:
                  type: string
    get:
      x-skaffolder-id: 5db331b4bcbe47264d820b7c
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5db331a8bcbe47264d820b72
      x-skaffolder-resource: Dipendente
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Dipendente
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dipendente"
    
  /dipendente/{id}:
    delete:
      x-skaffolder-id: 5db331b4bcbe47264d820b7d
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5db331a8bcbe47264d820b72
      x-skaffolder-resource: Dipendente
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Dipendente
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5db331b3bcbe47264d820b78
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5db331a8bcbe47264d820b72
      x-skaffolder-resource: Dipendente
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Dipendente
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dipendente"
    post:
      x-skaffolder-id: 5db331b3bcbe47264d820b77
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5db331a8bcbe47264d820b72
      x-skaffolder-resource: Dipendente
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Dipendente
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dipendente"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Dipendente"
    
  #-----  User  -----
  /user/{id}/changePassword:
    post:
      x-skaffolder-id: 5db330c0bcbe47264d820b44
      x-skaffolder-name: changePassword
      x-skaffolder-id-resource: 5db330bfbcbe47264d820b40
      x-skaffolder-resource: User
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: Change password of user from admin
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: object
      x-skaffolder-url: /{id}/changePassword
      x-skaffolder-roles: ["ADMIN"]
      summary: Change password of user from admin
      tags:
        - User
      responses:
        "200":
          description: OK
    
  /user/:
    post:
      x-skaffolder-id: 5db330c0bcbe47264d820b45
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5db330bfbcbe47264d820b40
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5db330c0bcbe47264d820b4a
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5db330bfbcbe47264d820b40
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    
  /user/{id}:
    delete:
      x-skaffolder-id: 5db330c0bcbe47264d820b46
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5db330bfbcbe47264d820b40
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5db330c0bcbe47264d820b43
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5db330bfbcbe47264d820b40
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5db330c0bcbe47264d820b49
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5db330bfbcbe47264d820b40
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
    
# JWT security
security:
  - bearerAuth: [] 
  
  
# Components
components:

  # Security
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
        

    Dipendente:
      x-skaffolder-id: 5db331a8bcbe47264d820b72
      x-skaffolder-id-db: 5db330bfbcbe47264d820b37
      x-skaffolder-id-entity: 5db331a8bcbe47264d820b70
      x-skaffolder-url: /dipendente
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Nome:
          type: string
          x-skaffolder-id-attr: 5db331a8bcbe47264d820b71
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
        Utenti:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5db331a8bcbe47264d820b70
          x-skaffolder-ent2: 5db330bfbcbe47264d820b38
      required:
        - _id
        - Nome
      
    
    User:
      x-skaffolder-id: 5db330bfbcbe47264d820b40
      x-skaffolder-id-db: 5db330bfbcbe47264d820b37
      x-skaffolder-id-entity: 5db330bfbcbe47264d820b38
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5db330bfbcbe47264d820b3b
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5db330bfbcbe47264d820b3c
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5db330bfbcbe47264d820b3a
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5db330bfbcbe47264d820b3e
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5db330bfbcbe47264d820b3d
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5db330bfbcbe47264d820b39
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
      
    
    
  # Databases
  x-skaffolder-db:
    - x-skaffolder-id: 5db330bfbcbe47264d820b37
      x-skaffolder-name: Demo1_db

  # Pages
  x-skaffolder-page:
    - x-skaffolder-id: 5db331b3bcbe47264d820b75
      x-skaffolder-name: DipendenteEdit
      x-skaffolder-url: /dipendentes/{id}
      x-skaffolder-template: Edit_Crud
      x-skaffolder-resource: 5db331a8bcbe47264d820b72
      x-skaffolder-services: 
        - 5db331b3bcbe47264d820b76
        - 5db331b3bcbe47264d820b78
        - 5db331b3bcbe47264d820b77
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
      x-skaffolder-roles: []
    - x-skaffolder-id: 5db331b3bcbe47264d820b7b
      x-skaffolder-name: DipendenteList
      x-skaffolder-url: /dipendentes
      x-skaffolder-template: List_Crud
      x-skaffolder-resource: 5db331a8bcbe47264d820b72
      x-skaffolder-services: 
        - 5db331b4bcbe47264d820b7d
        - 5db331b4bcbe47264d820b7c
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
        - 5db331b3bcbe47264d820b75
      x-skaffolder-roles: []
    - x-skaffolder-id: 5db330bfbcbe47264d820b3f
      x-skaffolder-name: Home
      x-skaffolder-url: /home
      x-skaffolder-template: 
      x-skaffolder-resource: 
      x-skaffolder-services: 
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
        - 5db331b3bcbe47264d820b7b
      x-skaffolder-roles: []