# src/main/java/com/demo1/controller/base/DipendenteBaseController.java
/* 
* Generated by
* 
*      _____ _          __  __      _     _
*     / ____| |        / _|/ _|    | |   | |
*    | (___ | | ____ _| |_| |_ ___ | | __| | ___ _ __
*     \___ \| |/ / _` |  _|  _/ _ \| |/ _` |/ _ \ '__|
*     ____) |   < (_| | | | || (_) | | (_| |  __/ |
*    |_____/|_|\_\__,_|_| |_| \___/|_|\__,_|\___|_|
*
* The code generator that works in many programming languages
*
*			https://www.skaffolder.com
*
*
* You can generate the code from the command-line
*       https://npmjs.com/package/skaffolder-cli
*
*       npm install -g skaffodler-cli
*
*   *   *   *   *   *   *   *   *   *   *   *   *   *   *   *
*
* To remove this comment please upgrade your plan here: 
*      https://app.skaffolder.com/#!/upgrade
*
* Or get up to 70% discount sharing your unique link:
*       https://app.skaffolder.com/#!/register?friend=5ad3007390b40a2b2d5f45fc
*
* You will get 10% discount for each one of your friends
* 
*/
package com.demo1.controller.base;

import java.util.List;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.bson.types.ObjectId;
import org.springframework.security.access.annotation.Secured;
import org.springframework.beans.factory.annotation.Autowired;
import com.demo1.db.demo1_db.service.DipendenteService;
import com.demo1.db.demo1_db.entity.Dipendente;

//IMPORT RELATIONS
import com.demo1.db.demo1_db.entity.User;

public class DipendenteBaseController {
    
    @Autowired
	DipendenteService dipendenteService;



//CRUD METHODS


    //CRUD - CREATE
    @Secured({ "ROLE_PRIVATE_USER" })
	@RequestMapping(value = "/dipendente",method = RequestMethod.POST, headers = "Accept=application/json")
	public Dipendente insert(@RequestBody Dipendente obj) {
		 return dipendenteService.insert(obj);
	}

	
    //CRUD - REMOVE
    @Secured({ "ROLE_PRIVATE_USER" })
	@RequestMapping(value = "/dipendente/{id}", method = RequestMethod.DELETE, headers = "Accept=application/json")
	public void delete(@PathVariable("id") String id) {
		dipendenteService.delete(id);
	}
	
	
    //CRUD - GET ONE
    @Secured({ "ROLE_PRIVATE_USER" })
	@RequestMapping(value = "/dipendente/{id}", method = RequestMethod.GET, headers = "Accept=application/json")
	public Dipendente get(@PathVariable String id) {
		return dipendenteService.get(id);
	}
	
	
    //CRUD - GET LIST
    @Secured({ "ROLE_PRIVATE_USER" })
	@RequestMapping(value = "/dipendente", method = RequestMethod.GET, headers = "Accept=application/json")
	public List<Dipendente> getList() {
		return dipendenteService.getList();
	}
	
	

    //CRUD - EDIT
    @Secured({ "ROLE_PRIVATE_USER" })
	@RequestMapping(value = "/dipendente/{id}", method = RequestMethod.POST, headers = "Accept=application/json")
	public Dipendente update(@RequestBody Dipendente obj) {
		return dipendenteService.update(obj);
	}
	


/*
 * CUSTOM SERVICES
 * 
 *	These services will be overwritten and implemented in  Custom.js
 */


	
}